user_manager/requirements.txt
flask
mysql-connector-python
grpcio
protobuf

